requirements.txt
python-dotenv==1.0.1
streamlit==1.41.1
openai==1.59.6
langsmith==0.2.11
